requirements.txt
streamlit>=1.24.0
numpy>=1.21.0
opencv-python>=4.7.0
Pillow>=9.4.0
ultralytics>=8.0.0
requests>=2.28.0
python-dotenv>=1.0.0 